Update method with tests, functional eager loading, unassign effect and assign effect

+++ OatMilk.Backend.Api/Services/Models/Requests/ModifierRequest.cs
@@ -7,7 +7,7 @@ namespace OatMilk.Backend.Api.Services.Models.Requests
     {
         public string Attribute { get; set; }
         
-        public float Magnitude { get; set; }
+        public double Magnitude { get; set; }
         
         public ModifierOperation Operation { get; set; }
     }
